refactor package directory
