+++ Rider/app/build.gradle
@@ -7,7 +7,7 @@ apply plugin: 'kotlin-android-extensions'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.example.rider"
+        applicationId "com.dti.rider"
         minSdkVersion 15
         targetSdkVersion 28
         versionCode 1
