+++ Rider/app/src/androidTest/java/com/dti/rider/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.rider
+package com.dti.rider
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
