+++ Rider/app/src/main/java/com/dti/rider/LoginActivity.kt
@@ -1,0 +1,38 @@
+package com.dti.rider
+
+import android.content.Intent
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.view.View
+import com.google.android.gms.auth.api.signin.GoogleSignIn
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount
+import com.google.android.gms.auth.api.signin.GoogleSignInClient
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions
+
+class LoginActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_login)
+        supportActionBar?.hide()
+
+
+        val gso : GoogleSignInOptions = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+            .requestEmail()
+            .build()
+
+        val mGoogleSignInClient : GoogleSignInClient = GoogleSignIn.getClient(this, gso)
+
+    }
+
+    override fun onStart() {
+        super.onStart()
+
+        val account : GoogleSignInAccount
+    }
+
+    fun signIn(view: View) {
+        val intent = Intent(this,MainActivity::class.java)
+        startActivity(intent)
+    }
+}
