+++ Rider/app/src/main/java/com/dti/rider/ProfileActivity.kt
@@ -1,0 +1,12 @@
+package com.dti.rider
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+
+class ProfileActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_profile)
+    }
+}
