+++ Rider/app/src/main/java/com/dti/rider/ThirdFragment.kt
@@ -1,4 +1,4 @@
-package com.example.rider
+package com.dti.rider
 
 
 import android.os.Bundle
@@ -10,14 +10,14 @@ import android.view.ViewGroup
 /**
  * A simple [Fragment] subclass.
  */
-class SecondFragment : Fragment() {
+class ThirdFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_second, container, false)
+        return inflater.inflate(R.layout.fragment_third, container, false)
     }
