--- Rider/app/src/main/java/com/example/rider/LoginActivity.kt
@@ -1,20 +1,0 @@
-package com.example.rider
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.view.View
-
-class LoginActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_login)
-        supportActionBar?.hide()
-    }
-
-    fun signIn(view: View) {
-        val intent = Intent(this,MainActivity::class.java)
-        startActivity(intent)
-    }
-}
